+++ App/Screens/UpdateSubCatagories.js
@@ -1,264 +1,230 @@
 import CheckBox from '@react-native-community/checkbox';
-import { Picker } from '@react-native-picker/picker';
 import axios from 'axios';
 import React, { useContext, useEffect, useState } from 'react';
-import { ActivityIndicator, Alert, Image, StyleSheet, Text, TextInput, TouchableOpacity, View } from 'react-native';
-import { commonStyle } from './../Config/DefaultCodes';
-import { AppContext } from './../Context/AppContext';
-import ModalForUser from './modal';
+import { FlatList, StatusBar, StyleSheet, Text, TextInput, TouchableOpacity, View } from 'react-native';
+import DropDownPicker from 'react-native-dropdown-picker';
+import { commonStyle } from '../Config/DefaultCodes';
+import { AppContext } from '../Context/AppContext';
+import SuccessModal from './SuccessModal';
+import ToggleDrawer from './ToggleDrawer';
 
-const UpdateSubCatagories = ({navigation}) => {
-    const { parentCatagoryData, setModal } = useContext(AppContext)
-    const [subCatagories, setSubCatagories] = useState('')
-    const [loading, setLoading] = useState(false)
-    const [textInput, setTextInput] = useState("")
-    const [toggleCheckBox, setToggleCheckBox] = useState(false)
+const UpdateSubCatagories = ({ navigation }) => {
+    const { parentCatagory, setParentCatagory, setSuccessModal } = useContext(AppContext)
+    const [dropDownOpen, setDropDownOpen] = useState(false);
+    const [selectedMainCatagory, setSelectedMainCatagory] = useState(null);
+    const [subCatagories, setSubCatagories] = useState('');
+    const [subCatagoriesWithData, setSubCatagoriesWithData] = useState('')
 
-    const [selectedCatagory, setSelectedCatagory] = useState();
-    const [seletedSubCatagory, setSeletedSubCatagory] = useState();
+    const [textInput, setTextInput] = useState(null);
+    const [toggleCheckBox, setToggleCheckBox] = useState(false);
+    const [defaultValue, setDefaultValue] = useState();
+    const [editModeActive, setEditModeActive] = useState(false);
 
+
+
+    // fetch-Sub-Catagories: 
     useEffect(() => {
-        if (selectedCatagory != null) {
-            const url = `https://myaccount.accountingarif.com/api/v1/account/sub-by-parent/${selectedCatagory}`
+        if (selectedMainCatagory != null) {
+            const url = `https://myaccount.accountingarif.com/api/v1/account/sub-by-parent/${selectedMainCatagory}`
             axios.get(url)
-                .then(res => setSubCatagories(res.data.data))
+                .then(res => {
+                    setSubCatagoriesWithData(res.data.data)
+                })
                 .catch(error => console.log(error))
         }
-    }, [selectedCatagory])
+    }, [selectedMainCatagory])
 
-    
+    useEffect(() => {
+        if (subCatagoriesWithData?.length > 0) {
+            setSubCatagories(
+                subCatagoriesWithData.map((data) => ({
+                    name: data.name,
+                    isActive: data.isActive,
+                    id: data.id,
+                    parentId: data.parentId,
+                }))
+            )
+        } else {
+            setSubCatagories()
+        }
+    }, [subCatagoriesWithData])
+
+
+    const editTheSubCatagory = (id, name, parentId) => {
+        const update = {
+            oldName: name,
+            catagoryId: id,
+            parentId: parentId,
+        }
+        setEditModeActive(true)
+        setDefaultValue(update)
+    }
+
     const updateCatagory = () => {
 
-        const name = textInput.name
-        if (seletedSubCatagory == null) {
-            Alert.alert("Catagory Null", "Please select Catagory")
-        } else if (name == "") {
-            Alert.alert("No name", "Please write a name")
-        } else {
-            setLoading(true)
-            const dataToSent = new FormData();
-            dataToSent.append("id", seletedSubCatagory);
-            dataToSent.append("name", textInput.name);
-            dataToSent.append("icon", null)
-            dataToSent.append("isParent", false);
-            dataToSent.append("parentId", selectedCatagory);
-            dataToSent.append("isCash", false);
-            dataToSent.append("isActive", toggleCheckBox);
-            console.log(dataToSent);
-            axios.post("https://myaccount.accountingarif.com/api/v1/account/update", dataToSent)
-                .then(res => {
-                    console.log("server Res :", res.data)
-                    console.log("res has reached.")
-                    if (res.data.success == true) {
-                        setModal(true)
-                    }
-                })
-                .catch(error => console.log("Server Errr :", error))
-        }
-        console.log("Cliked")
+        const dataToSent = new FormData();
+        dataToSent.append("id", defaultValue.catagoryId);
+        dataToSent.append("name", textInput);
+        dataToSent.append("icon", null)
+        dataToSent.append("isParent", false);
+        dataToSent.append("parentId", defaultValue.parentId);
+        dataToSent.append("isCash", false);
+        dataToSent.append("isActive", toggleCheckBox);
+        console.log(dataToSent);
+        axios.post("https://myaccount.accountingarif.com/api/v1/account/update", dataToSent)
+            .then(res => {
+                console.log("server Res :", res.data)
+                if (res.data.success == true) {
+                    setSuccessModal(true)
+                }
+            })
+            .catch(error => console.log("Server Errr :", error))
+        setDefaultValue()
+        setEditModeActive(false)
     }
-    useEffect(() => {
-        setTimeout(() => {
-            setLoading(false)
-        }, 1000)
-    }, [loading])
 
-    if (loading) {
-        return (
-            <View style={{ flex: 1, alignItems: "center", justifyContent: "center" }}>
-                <ActivityIndicator size="large" color="grey" />
-            </View>
-        )
-    }
+
     return (
         <View style={commonStyle.container}>
-             <View style={{
-                position: "absolute",
-                left: 10,
-                top: 20,
-            }}>
-                <TouchableOpacity onPress={() => navigation.toggleDrawer()}>
-                    <Image
-                        style={{
-                            height: 30,
-                            width: 30,
-                        }}
-                        source={require("../assets/icons/menu.png")}
-                    />
-                </TouchableOpacity>
-            </View>
-            <ModalForUser>
-                <View>
-                    <TouchableOpacity onPress={() => setModal(false)} style={{
-                        width: '100%',
-                        height: 20,
-                        alignItems: 'flex-end',
-                        justifyContent: 'center',
-                    }}>
-                        <Image
-                            style={{
-                                width: 30,
-                                height: 30,
-                            }}
-                            source={require("../assets/icons/modal/x.png")} />
-                    </TouchableOpacity>
-                    <Image
-                        style={{
-                            height: 150,
-                            width: 150,
-                        }}
-                        source={require("../assets/icons/modal/checked.png")} />
-                </View>
-            </ModalForUser>
-            <View style={{ alignItems: "center", marginVertical: 20, }}>
-                <Text style={{ fontSize: 25, }}>Update sub catagory</Text>
-                <View style={styles.addSubCatagories}>
-                    <View style={{
-                        borderRadius: 10,
-                        padding: 5,
-                        borderColor: "#DDDDDD",
-                        borderWidth: 2,
-                        marginBottom: 10,
-                    }}>
-                        <Text style={{ fontSize: 18, marginLeft: 5, }}>Main catagory</Text>
-                        <Picker
-                            style={{
-                                height: 40,
-                                marginTop: 10,
-                                marginBottom: 10,
-                            }}
-                            dropdownIconColor="#5C527F"
-                            mode="dropdown"
-                            selectedValue={selectedCatagory}
-                            onValueChange={(itemValue, itemIndex) =>
-                                setSelectedCatagory(itemValue)
-                            }>
-                            <Picker.Item label='Please select catagory' value={null} />
-                            {
-                                parentCatagoryData.map(data => {
-                                    return <Picker.Item
-                                        key={data.id}
-                                        label={data.name}
-                                        value={data.id}
-                                        enabled={true}
-                                        style={{
-                                            color: "#5C527F",
-                                            fontSize: 22,
-                                        }} />
-                                })
-                            }
-                        </Picker>
-                    </View>
+             <StatusBar
+                barStyle="dark-content"
+                backgroundColor="#fff"
+                translucent={true}
+            />
+            <SuccessModal />
+            <ToggleDrawer
+                navigation={navigation}
+            />
+            <View style={commonStyle.formContainer}>
+                <View style={commonStyle.formWrapper}>
                     <View>
-                        {subCatagories ? <View style={{
-                            borderRadius: 10,
-                            padding: 5,
-                            borderColor: "#DDDDDD",
-                            borderWidth: 2,
-                            marginBottom: 10,
-
-                        }}>
-                            <Text style={{
-                                fontSize: 18,
-                                marginLeft: 5,
-                            }}>Sub catagory which will update</Text>
-                            <Picker
-                                style={{
-                                    height: 40,
-                                    marginTop: 10,
-                                    marginBottom: 10,
-                                }}
-                                dropdownIconColor="#5C527F"
-                                mode="dropdown"
-                                selectedValue={seletedSubCatagory}
-                                onValueChange={(itemValue, itemIndex) =>
-                                    setSeletedSubCatagory(itemValue)
-                                }>
-                                <Picker.Item label='Please select sub catagory' value={null} />
-
-                                {
-                                    subCatagories.map(data => {
-                                        return (
-                                            <Picker.Item
-                                                key={data.id}
-                                                label={data.name}
-                                                value={data.id}
-                                                enabled={true}
-                                                style={{
-                                                    color: "#5C527F",
-                                                    fontSize: 22,
-                                                }} />
-                                        )
-                                    })
-                                }
-                            </Picker>
-                        </View> : null}
-                    </View>
-                    <TextInput
-                        style={styles.textInput}
-                        placeholder="Sub catagory name"
-                        value={textInput.name}
-                        onChangeText={text => setTextInput({ ...textInput, name: text })}
-                    />
-
-                    <View style={styles.checkBox}>
-                        <CheckBox
-                            disabled={false}
-                            tintColors={{ true: '#F54748', false: 'black' }}
-                            value={toggleCheckBox}
-                            onValueChange={(newValue) => setToggleCheckBox(newValue)}
-                        />
-                        <Text style={{ fontSize: 16 }}>Active this sub catagory</Text>
-                    </View>
-
-                    <TouchableOpacity style={styles.submitBtn} onPress={() => updateCatagory()}>
-                        <Text style={{ color: "#fff", fontSize: 18 }}>
-                            Update
+                        <Text style={commonStyle.pageTitle}>
+                            Update sub catagories
                         </Text>
-                    </TouchableOpacity>
+                    </View>
+                    <View style={{ alignItems: "center" }}>
+                        <View style={[commonStyle.dropDownContainer, dropDownOpen ? { minHeight: 260 } : null]}>
+                            <DropDownPicker
+                                placeholderStyle={{
+                                    color: "#5C527F",
+                                    fontWeight: "bold"
+                                }}
+                                textStyle={{
+                                    fontSize: 18
+                                }}
+                                labelStyle={{
+                                    fontWeight: "bold"
+                                }}
+                                style={{
+                                    borderColor: "#DDDDDD",
+                                    borderWidth: 2,
+                                }}
+                                dropDownContainerStyle={{
+                                    zIndex: 1000,
+                                    elevation: 1000,
+                                    marginTop: 5,
+                                    paddingHorizontal: 10,
+                                    paddingVertical: 10,
+                                    borderWidth: 3,
+                                    borderRadius: 20,
+                                    borderColor: "#DDDDDD",
+                                }}
+                                selectedItemContainerStyle={{
+                                    backgroundColor: "#79B4B7",
+                                    borderRadius: 10,
+                                    height: 50,
+                                }}
+                                placeholder="Select catagory"
+                                open={dropDownOpen}
+                                value={selectedMainCatagory}
+                                items={parentCatagory}
+                                setOpen={setDropDownOpen}
+                                setValue={setSelectedMainCatagory}
+                                setItems={setParentCatagory}
+                            />
+                        </View>
+                    </View>
 
+                    {editModeActive ?
+                        <View style={{ marginTop: 10, alignItems: "center" }}>
+
+                            <TextInput
+                                style={commonStyle.textInput}
+                                defaultValue={defaultValue.oldName}
+                                value={textInput}
+                                onChangeText={text => setTextInput(text)}
+                            />
+
+                            <View style={commonStyle.checkBox}>
+                                <CheckBox
+                                    disabled={false}
+                                    tintColors={{ true: '#F54748', false: 'black' }}
+                                    value={toggleCheckBox}
+                                    onValueChange={(newValue) => setToggleCheckBox(newValue)}
+                                />
+                                <Text style={{ fontSize: 18, }}>Active</Text>
+                            </View>
+                            <TouchableOpacity
+                                style={[commonStyle.submitORsaveBtn, { width: "80%" }]}
+                                onPress={() => updateCatagory()}>
+                                <Text style={{ color: "#fff", fontSize: 18 }}>
+                                    Update
+                                </Text>
+                            </TouchableOpacity>
+                        </View>
+                        : null}
+
+                    <View style={{
+                        maxHeight: 500,
+                        width: "95%",
+                        marginHorizontal: 10,
+                    }}>
+                        <Text style={{ textAlign: "center", marginTop: 10, fontSize: 17 }}>Available sub catagories</Text>
+                        {subCatagories ? <FlatList
+
+                            style={{
+                                marginTop: 10,
+                                paddingVertical: 10,
+                                paddingHorizontal: 10,
+                                borderWidth: 3,
+                                borderRadius: 10,
+                                borderColor: "#DDDDDD",
+                            }}
+                            data={subCatagories}
+                            renderItem={({ item }) => (
+                                <View style={{
+                                    flexDirection: "row",
+                                    justifyContent: "space-between",
+                                    marginBottom: 5,
+                                    marginTop: 10,
+                                }}>
+                                    <View>
+                                        <Text style={{
+                                            fontSize: 20,
+                                        }}>{item.name}</Text>
+                                    </View>
+
+                                    <TouchableOpacity style={commonStyle.smallButton}
+                                        onPress={() => editTheSubCatagory(item.id, item.name, item.parentId)}
+                                    >
+                                        <Text style={{ color: "#fff" }}>Edit</Text>
+                                    </TouchableOpacity>
+
+                                </View>
+                            )}
+                            keyExtractor={(item, index) => index.toString()}
+                        /> : null}
+                    </View>
                 </View>
             </View>
-        </View>
+        </View >
     )
 }
 
-export default UpdateSubCatagories
-
+export default UpdateSubCatagories;
 
 const styles = StyleSheet.create({
-    // container: {
-    //     flex: 1,
-    //     marginTop: StatusBar.currentHeight,
-    // },
-    addSubCatagories: {
-        // marginHorizontal: 10,
-        padding: 10,
-        width: "80%",
-    },
-    textInput: {
-        width:"80%",
-        height: 60,
-        marginBottom: 10,
-        paddingLeft: 10,
-        fontSize: 18,
-        borderWidth: 2,
-        borderColor: "#DDDDDD",
-        borderRadius: 10,
-        backgroundColor: "#fff",
-    },
-    checkBox: {
-        flexDirection: "row",
-        alignItems: "center",
-        fontSize: 20,
-        marginTop: 10,
-    },
-    submitBtn: {
-        alignItems: "center",
-        justifyContent: "center",
-        height: 40,
-        marginTop: 10,
-        borderRadius: 10,
-        backgroundColor: "#8236CB",
-    },
+    // formWrapper: {
+    //     marginTop: 30,
+    // }
 })